refactor : throw 대신 null 값을 통해 올바른 로직 구현
