+++ src/main/java/com/sparta/deliverybackend/domain/order/controller/OrderController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.sparta.deliverybackend.api.controller.dto.VerifiedMember;
+import com.sparta.deliverybackend.api.auth.controller.dto.VerifiedMember;
 import com.sparta.deliverybackend.domain.order.controller.dto.OrderCancelResponseDto;
 import com.sparta.deliverybackend.domain.order.controller.dto.OrderMenuDto;
 import com.sparta.deliverybackend.domain.order.controller.dto.OrderRequestDto;
