+++ src/main/java/com/sparta/deliverybackend/domain/restaurant/controller/MenuController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.sparta.deliverybackend.api.controller.dto.VerifiedMember;
+import com.sparta.deliverybackend.api.auth.controller.dto.VerifiedMember;
 import com.sparta.deliverybackend.domain.restaurant.controller.dto.MenuCreateReqDto;
 import com.sparta.deliverybackend.domain.restaurant.controller.dto.MenuRespDto;
 import com.sparta.deliverybackend.domain.restaurant.controller.dto.MenuUpdateReqDto;
@@ -28,7 +28,8 @@ public class MenuController {
 
 	// 메뉴 생성
 	@PostMapping()
-	public ResponseEntity<MenuRespDto> createMenu(@RequestBody MenuCreateReqDto menuCreateReqDto, VerifiedMember verifiedMember){
+	public ResponseEntity<MenuRespDto> createMenu(@RequestBody MenuCreateReqDto menuCreateReqDto,
+		VerifiedMember verifiedMember) {
 		MenuRespDto menuRespDto = menuService.createMenu(menuCreateReqDto, verifiedMember);
 		return ResponseEntity
 			.status(HttpStatus.CREATED)
@@ -37,8 +38,8 @@ public class MenuController {
 
 	// 메뉴 수정
 	@PatchMapping("/{restaurantId}/menus/{menuId}")
-	public ResponseEntity<Void> updateMenu(@PathVariable Long restaurantId,	@PathVariable Long menuId,
-		@RequestBody @Valid MenuUpdateReqDto menuUpdateReqDto, VerifiedMember verifiedMember){
+	public ResponseEntity<Void> updateMenu(@PathVariable Long restaurantId, @PathVariable Long menuId,
+		@RequestBody @Valid MenuUpdateReqDto menuUpdateReqDto, VerifiedMember verifiedMember) {
 		menuService.updateMenu(restaurantId, menuId, menuUpdateReqDto, verifiedMember);
 		return ResponseEntity
 			.status(HttpStatus.NO_CONTENT)
@@ -47,7 +48,8 @@ public class MenuController {
 
 	// 메뉴 삭제
 	@DeleteMapping("/{restaurantId}/menus/{menuId}")
-	public ResponseEntity<Void> deleteMenu(@PathVariable Long restaurantId, @PathVariable Long menuId, VerifiedMember verifiedMember){
+	public ResponseEntity<Void> deleteMenu(@PathVariable Long restaurantId, @PathVariable Long menuId,
+		VerifiedMember verifiedMember) {
 		menuService.deleteMenu(restaurantId, menuId, verifiedMember);
 		return ResponseEntity
 			.status(HttpStatus.NO_CONTENT)
